mega refactor to using GTFS as data source - keep same functionality but enhacing Renfe python class - preparing new major release
