+++ src/cli.rs
@@ -3,10 +3,7 @@ use getopts::Options;
 use pyo3::{exceptions::PyValueError, pyfunction, PyResult};
 use std::env;
 
-use crate::{
-    stations::Renfe,
-    timetable::{print_timetable, search_timetable},
-};
+use crate::renfe::Renfe;
 
 #[pyfunction]
 pub fn main() -> PyResult<()> {
@@ -14,7 +11,6 @@ pub fn main() -> PyResult<()> {
     let program = args[0].clone();
     let now = Utc::now();
     let opts = set_opts();
-    let renfe = Renfe::new()?;
 
     let matches = match opts.parse(&args[1..]) {
         Ok(m) => m,
@@ -27,37 +23,42 @@ pub fn main() -> PyResult<()> {
         print_usage(&program, opts);
         return Ok(());
     }
+
+    let mut renfe = Renfe::new()?;
+
     let origin = renfe.filter_station(matches.opt_str("f").unwrap_or("".to_owned()))?;
     let destination = renfe.filter_station(matches.opt_str("t").unwrap_or("".to_owned()))?;
-    let day = enrich_day(matches.opt_str("d").unwrap_or(now.day().to_string()));
-    let month = matches.opt_str("m").unwrap_or(now.month().to_string());
-    let year = matches.opt_str("y").unwrap_or(now.year().to_string());
-    let wait = matches
-        .opt_str("w")
-        .unwrap_or(2.to_string())
-        .parse::<u64>()?;
+    let day = match matches.opt_str("d") {
+        Some(day) => day.parse().unwrap(),
+        None => now.day(),
+    };
+    let month = match matches.opt_str("m") {
+        Some(day) => day.parse().unwrap(),
+        None => now.month(),
+    };    
+    let year = match matches.opt_str("y") {
+        Some(day) => day.parse().unwrap(),
+        None => now.year(),
+    };
     let sorted: bool = matches.opt_present("s");
 
-    println!(
-        "Today is: {}-{}-{}",
-        now.year(),
-        now.month(),
-        enrich_day(now.day().to_string())
-    );
+    println!("Today is: {}-{}-{}", now.year(), now.month(), now.day());
     println!("Searching timetable for date: {}-{}-{}", year, month, day);
 
-    let timetable = search_timetable(origin, destination, day, month, year, wait, sorted)?;
+    renfe.set_train_schedules(&origin.1, &destination.1, day, month, year, sorted)?;
 
-    print_timetable(timetable);
+    println!(
+        "Origin station: {}",
+        origin.0
+    );
+    println!(
+        "Destination station: {}",
+        destination.0
+    );
+
+    renfe.print_timetable();
+
     Ok(())
-}
-
-fn enrich_day(day: String) -> String {
-    if day.len() == 1 {
-        "0".to_owned() + &day
-    } else {
-        day
-    }
 }
 
 fn print_usage(program: &str, opts: Options) {
@@ -86,12 +87,6 @@ fn set_opts() -> Options {
         "year",
         "Set Year to search timetable for (default: today's year)",
         "YEAR",
-    );
-    opts.optopt(
-        "w",
-        "wait",
-        "Set Wait time in seconds for Renfe search result page (default: 2)",
-        "SECONDS",
     );
     opts.optflag("s", "sort", "Option to sort the timetable by Duration");
     opts.optflag("h", "help", "Print this help menu");
